Components integrated with Interface to remove direct calls
