+++ src/main/scala/de/htwg/se/wizard/model/roundComponent/RoundInterface.scala
@@ -1,0 +1,5 @@
+package de.htwg.se.wizard.model.roundComponent
+
+trait RoundInterface {
+  def madeTricks(list: Iterable[Int]): Round
+}
