+++ src/main/scala/de/htwg/se/wizard/util/Command.scala
@@ -1,0 +1,7 @@
+package de.htwg.se.wizard.util
+
+trait Command {
+  def doStep():Unit
+  def undoStep():Unit
+  def redoStep():Unit
+}
