+++ src/test/scala/de/htwg/se/wizard/model/RandomTest.scala
@@ -1,4 +1,5 @@
 package de.htwg.se.wizard.model
+import de.htwg.se.wizard.model.cardsComponent.Cards
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
