+++ src/test/scala/de/htwg/se/wizard/model/RoundTest.scala
@@ -1,4 +1,5 @@
 package de.htwg.se.wizard.model
+import de.htwg.se.wizard.model.roundComponent.Round
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
